refactor(server): migrate core backend modules from CommonJS to ESM
