+++ server/utils/logger.js
@@ -3,11 +3,11 @@
  * Structured logging for all backend operations
  */
 
-const winston = require("winston");
-const path = require("path");
+import winston from "winston";
+import path from "path";
 
 // Create logs directory if it doesn't exist
-const fs = require("fs");
+import fs from "fs";
 const logsDir = path.join(process.cwd(), "logs");
 if (!fs.existsSync(logsDir)) {
   fs.mkdirSync(logsDir, { recursive: true });
@@ -99,4 +99,4 @@ const logger = winston.createLogger({
   ],
 });
 
-module.exports = logger;
+export default logger;
